# src/DirectiveDrift.Web/src/api/generated/types.gen.ts
// @ts-nocheck -- generated transport is checked by API drift tests
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'http://127.0.0.1:5078/' | (string & {});
};

export type AnonymousTypeOfstringAndstringAndstring = {
    missionId: null | string;
    title: null | string;
    contentVersion: null | string;
};

export type JsonElement = unknown;

export type RuntimeResponse = {
    apiVersion: string;
    providerMode: string;
    stateSchemaVersion: string;
};

export type StartRunRequest = {
    buildId: string;
    buildVersion: number | string;
    variantId: string;
};

export type GetApiV1RuntimeData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/runtime';
};

export type GetApiV1RuntimeResponses = {
    /**
     * OK
     */
    200: RuntimeResponse;
};

export type GetApiV1RuntimeResponse = GetApiV1RuntimeResponses[keyof GetApiV1RuntimeResponses];

export type GetApiV1MissionsData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/missions';
};

export type GetApiV1MissionsResponses = {
    /**
     * OK
     */
    200: Array<AnonymousTypeOfstringAndstringAndstring>;
};

export type GetApiV1MissionsResponse = GetApiV1MissionsResponses[keyof GetApiV1MissionsResponses];

export type GetApiV1MissionsByMissionIdData = {
    body?: never;
    path: {
        missionId: string;
    };
    query?: never;
    url: '/api/v1/missions/{missionId}';
};

export type GetApiV1MissionsByMissionIdResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type GetApiV1MissionsByMissionIdPracticeVariantsData = {
    body?: never;
    path: {
        missionId: string;
    };
    query?: never;
    url: '/api/v1/missions/{missionId}/practice-variants';
};

export type GetApiV1MissionsByMissionIdPracticeVariantsResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type GetApiV1BuildsData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/builds';
};

export type GetApiV1BuildsResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type PostApiV1BuildsData = {
    body: JsonElement;
    path?: never;
    query?: never;
    url: '/api/v1/builds';
};

export type PostApiV1BuildsResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type GetApiV1BuildsByBuildIdData = {
    body?: never;
    path: {
        buildId: string;
    };
    query?: never;
    url: '/api/v1/builds/{buildId}';
};

export type GetApiV1BuildsByBuildIdResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type GetApiV1BuildsByBuildIdVersionsData = {
    body?: never;
    path: {
        buildId: string;
    };
    query?: never;
    url: '/api/v1/builds/{buildId}/versions';
};

export type GetApiV1BuildsByBuildIdVersionsResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type PostApiV1BuildsByBuildIdVersionsData = {
    body: JsonElement;
    path: {
        buildId: string;
    };
    query?: never;
    url: '/api/v1/builds/{buildId}/versions';
};

export type PostApiV1BuildsByBuildIdVersionsResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type PostApiV1RunsData = {
    body: StartRunRequest;
    path?: never;
    query?: never;
    url: '/api/v1/runs';
};

export type PostApiV1RunsResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type GetApiV1RunsByRunIdData = {
    body?: never;
    path: {
        runId: string;
    };
    query?: never;
    url: '/api/v1/runs/{runId}';
};

export type GetApiV1RunsByRunIdResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type PostApiV1RunsByRunIdTurnsData = {
    body?: never;
    path: {
        runId: string;
    };
    query?: never;
    url: '/api/v1/runs/{runId}/turns';
};

export type PostApiV1RunsByRunIdTurnsResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type GetApiV1OperationsByOperationIdData = {
    body?: never;
    path: {
        operationId: string;
    };
    query?: never;
    url: '/api/v1/operations/{operationId}';
};

export type GetApiV1OperationsByOperationIdResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type GetApiV1RunsByRunIdEventsData = {
    body?: never;
    path: {
        runId: string;
    };
    query?: {
        afterSequence?: number | string;
        limit?: number | string;
    };
    url: '/api/v1/runs/{runId}/events';
};

export type GetApiV1RunsByRunIdEventsResponses = {
    /**
     * OK
     */
    200: unknown;
};

export type GetApiV1RunsByRunIdReplayData = {
    body?: never;
    path: {
        runId: string;
    };
    query?: never;
    url: '/api/v1/runs/{runId}/replay';
};

export type GetApiV1RunsByRunIdReplayResponses = {
    /**
     * OK
     */
    200: unknown;
};